Update go.mod and go.sum files
This commit updates the `go.mod` and `go.sum` files to reflect the changes in package versions. In `go.mod`, the Go version requirement is updated from `1.21` to `1.22`. The `github.com/PuerkitoBio/goquery` package version is updated from `v1.8.1` to `v1.9.2`. The `github.com/andybalholm/cascadia` package version is updated from `v1.3.1` to `v1.3.2`. The `golang.org/x/net` package version is updated from `v0.17.0` to `v0.25.0`. In `go.sum`, the checksums for the updated package versions are adjusted accordingly.